Refactor inference functions; run_tf_detector.py no longer self-contained.
