+++ detection/run_tf_detector.py
@@ -1,14 +1,12 @@
 """
 Module to run a TensorFlow animal detection model on images.
 
-Contains functions to load a TensorFlow detection model, run inference,
-render bounding boxes on images, and write out the resulting
-images (with bounding boxes).
+The class TFDetector contains functions to load a TensorFlow detection model and run inference.
+The main function in this script also render the predicted bounding boxes on images and
+save the resulting images (with bounding boxes).
 
-THIS SCRIPT IS NOT A GOOD WAY TO PROCESS LOTS OF IMAGES.
-
-Did we mention in all caps that this script is not a good way to process lots
-of images? It does not facilitate checkpointing the results so if it crashes
+This script is not a good way to process lots of images (tens of thousands, say).
+It does not facilitate checkpointing the results so if it crashes
 you would have to start from scratch. If you want to run a detector (e.g. ours)
 on lots of images, you should check out:
 
@@ -17,24 +15,20 @@ on lots of images, you should check out:
 2) https://github.com/microsoft/CameraTraps/tree/master/api/batch_processing
    (for running large jobs on Azure ML)
 
-The good news: this script depends on nothing else in our repo, just standard pip
-installs (the list of packages you need to install via pip is listed in `envrionment-detector.yml`
-at the root of the repo); alternatively, we recommend you set up a conda virtual environment
-following the Installation section on the main README, using `envrionment-detector.yml` as the
-environment file where asked).
+To run this script, we recommend you set up a conda virtual environment following instructions
+in the Installation section on the main README, using `environment-detector.yml` as the
+environment file where asked.
 
-It's a good way to test our detector on a handful of images and
+This is a good way to test our detector on a handful of images and
 get super-satisfying, graphical results.  It's also a good way to see how
 fast a detector model will run on a particular machine.
 
-See the "command-line driver" cell for example invocations.
-
-If you would like to not use the GPU on the machine, set the environment variable CUDA_VISIBLE_DEVICES to "-1"
+If you would like to *not* use the GPU on the machine, set the environment variable CUDA_VISIBLE_DEVICES to "-1"
 
 If no output directory is specified, writes detections for c:\foo\bar.jpg to
 c:\foo\bar_detections.jpg .
 
-This script will only consider detections with > 0.3 confidence at all times. The `threshold` you
+This script will only consider detections with > 0.1 confidence at all times. The `threshold` you
 provide is only for rendering the results. If you need to see lower-confidence detections, you can change
 DEFAULT_OUTPUT_CONFIDENCE_THRESHOLD.
 
@@ -47,7 +41,6 @@ https://github.com/tensorflow/models/blob/master/research/object_detection/infer
 
 import argparse
 import glob
-import math
 import os
 import statistics
 import sys
@@ -55,9 +48,11 @@ import time
 import warnings
 
 import humanfriendly
-from PIL import Image, ImageFont, ImageDraw
 import numpy as np
 from tqdm import tqdm
+
+from ct_utils import truncate_float
+import visualization.visualization_utils as viz_utils
 
 # ignoring all "PIL cannot read EXIF metainfo for the images" warnings
 warnings.filterwarnings('ignore', '(Possibly )?corrupt EXIF data', UserWarning)
@@ -115,361 +110,6 @@ class ImagePathUtils:
         return image_strings
 
 
-class DetectorUtils:
-    """
-    A collection of utility functions to support TFDetector.
-    The functions themselves live in ct_utils.py and visualization/visualization_utils.py
-    and should be imported from there instead - this script needs to be stand-alone.
-    """
-
-    COLORS = [
-        'AliceBlue', 'Red', 'RoyalBlue', 'Gold', 'Chartreuse', 'Aqua', 'Azure',
-        'Beige', 'Bisque', 'BlanchedAlmond', 'BlueViolet', 'BurlyWood', 'CadetBlue',
-        'AntiqueWhite', 'Chocolate', 'Coral', 'CornflowerBlue', 'Cornsilk', 'Crimson',
-        'Cyan', 'DarkCyan', 'DarkGoldenRod', 'DarkGrey', 'DarkKhaki', 'DarkOrange',
-        'DarkOrchid', 'DarkSalmon', 'DarkSeaGreen', 'DarkTurquoise', 'DarkViolet',
-        'DeepPink', 'DeepSkyBlue', 'DodgerBlue', 'FireBrick', 'FloralWhite',
-        'ForestGreen', 'Fuchsia', 'Gainsboro', 'GhostWhite', 'GoldenRod',
-        'Salmon', 'Tan', 'HoneyDew', 'HotPink', 'IndianRed', 'Ivory', 'Khaki',
-        'Lavender', 'LavenderBlush', 'LawnGreen', 'LemonChiffon', 'LightBlue',
-        'LightCoral', 'LightCyan', 'LightGoldenRodYellow', 'LightGray', 'LightGrey',
-        'LightGreen', 'LightPink', 'LightSalmon', 'LightSeaGreen', 'LightSkyBlue',
-        'LightSlateGray', 'LightSlateGrey', 'LightSteelBlue', 'LightYellow', 'Lime',
-        'LimeGreen', 'Linen', 'Magenta', 'MediumAquaMarine', 'MediumOrchid',
-        'MediumPurple', 'MediumSeaGreen', 'MediumSlateBlue', 'MediumSpringGreen',
-        'MediumTurquoise', 'MediumVioletRed', 'MintCream', 'MistyRose', 'Moccasin',
-        'NavajoWhite', 'OldLace', 'Olive', 'OliveDrab', 'Orange', 'OrangeRed',
-        'Orchid', 'PaleGoldenRod', 'PaleGreen', 'PaleTurquoise', 'PaleVioletRed',
-        'PapayaWhip', 'PeachPuff', 'Peru', 'Pink', 'Plum', 'PowderBlue', 'Purple',
-        'RosyBrown', 'Aquamarine', 'SaddleBrown', 'Green', 'SandyBrown',
-        'SeaGreen', 'SeaShell', 'Sienna', 'Silver', 'SkyBlue', 'SlateBlue',
-        'SlateGray', 'SlateGrey', 'Snow', 'SpringGreen', 'SteelBlue', 'GreenYellow',
-        'Teal', 'Thistle', 'Tomato', 'Turquoise', 'Violet', 'Wheat', 'White',
-        'WhiteSmoke', 'Yellow', 'YellowGreen'
-    ]
-
-    @staticmethod
-    def truncate_float(x, precision=3):
-        """
-        Function for truncating a float scalar to the defined precision.
-        For example: truncate_float(0.0003214884) --> 0.000321
-
-        Args:
-        x         (float) Scalar to truncate
-        precision (int)   The number of significant digits to preserve, should be
-                          greater or equal 1
-        """
-
-        assert precision > 0
-
-        if np.isclose(x, 0):
-            return 0
-        else:
-            # Determine the factor, which shifts the decimal point of x
-            # just behind the last significant digit
-            factor = math.pow(10, precision - 1 - math.floor(math.log10(abs(x))))
-            # Shift decimal point by multiplicatipon with factor, flooring, and
-            # division by factor
-            return math.floor(x * factor) / factor
-
-    @staticmethod
-    def round_and_make_float(d, precision=4):
-        return DetectorUtils.truncate_float(float(d), precision=precision)
-
-    @staticmethod
-    def __open_image(input_file):
-        """Opens an image in binary format using PIL.Image and convert to RGB mode.
-
-        Args:
-            input_file: an image in binary format read from the POST request's body or
-                path to an image file (anything that PIL can open)
-
-        Returns:
-            an PIL image object in RGB mode
-        """
-        image = Image.open(input_file)
-        if image.mode not in ('RGBA', 'RGB'):
-            raise AttributeError('Input image not in RGBA or RGB mode and cannot be processed.')
-        if image.mode == 'RGBA':
-            # PIL.Image.convert() returns a converted copy of this image
-            image = image.convert(mode='RGB')
-        return image
-
-    @staticmethod
-    def load_image(input_file):
-        """Loads the image at input_file as a PIL Image into memory; Image.open() is lazy and errors will
-        occur downstream if not explicitly loaded
-
-        Args:
-            input_file: an image in binary format read from the POST request's body or
-                path to an image file (anything that PIL can open)
-
-        Returns:
-            an PIL image object in RGB mode
-        """
-        image = DetectorUtils.__open_image(input_file)
-        image.load()
-        return image
-
-    # The following functions are modified versions of those at:
-    # https://github.com/tensorflow/models/blob/master/research/object_detection/utils/visualization_utils.py
-
-    @staticmethod
-    def render_detection_bounding_boxes(detections, image,
-                                        label_map={},
-                                        classification_label_map={},
-                                        confidence_threshold=0.8, thickness=4, expansion=0,
-                                        classification_confidence_threshold=0.3,
-                                        max_classifications=3):
-        """
-        Renders bounding boxes, label, and confidence on an image if confidence is above the threshold.
-
-        This works with the output of the batch processing API.
-
-        Supports classification, if the detection contains classification results according to the
-        API output version 1.0.
-
-        Args:
-
-            detections: detections on the image, example content:
-                [
-                    {
-                        "category": "2",
-                        "conf": 0.996,
-                        "bbox": [
-                            0.0,
-                            0.2762,
-                            0.1234,
-                            0.2458
-                        ]
-                    }
-                ]
-
-                ...where the bbox coordinates are [x, y, box_width, box_height].
-
-                (0, 0) is the upper-left.  Coordinates are normalized.
-
-                Supports classification results, if *detections* have the format
-                [
-                    {
-                        "category": "2",
-                        "conf": 0.996,
-                        "bbox": [
-                            0.0,
-                            0.2762,
-                            0.1234,
-                            0.2458
-                        ]
-                        "classifications": [
-                            ["3", 0.901],
-                            ["1", 0.071],
-                            ["4", 0.025]
-                        ]
-                    }
-                ]
-
-            image: PIL.Image object, output of generate_detections.
-
-            label_map: optional, mapping the numerical label to a string name. The type of the numerical label
-                (default string) needs to be consistent with the keys in label_map; no casting is carried out.
-
-            classification_label_map: optional, mapping of the string class labels to the actual class names.
-                The type of the numerical label (default string) needs to be consistent with the keys in
-                label_map; no casting is carried out.
-
-            confidence_threshold: optional, threshold above which the bounding box is rendered.
-            thickness: line thickness in pixels. Default value is 4.
-            expansion: number of pixels to expand bounding boxes on each side.  Default is 0.
-            classification_confidence_threshold: confidence above which classification result is retained.
-            max_classifications: maximum number of classification results retained for one image.
-
-        image is modified in place.
-        """
-
-        display_boxes = []
-        display_strs = []  # list of lists, one list of strings for each bounding box (to accommodate multiple labels)
-        classes = []  # for color selection
-
-        for detection in detections:
-
-            score = detection['conf']
-            if score >= confidence_threshold:
-
-                x1, y1, w_box, h_box = detection['bbox']
-                display_boxes.append([y1, x1, y1 + h_box, x1 + w_box])
-                clss = detection['category']
-                label = label_map[clss] if clss in label_map else clss
-                displayed_label = ['{}: {}%'.format(label, round(100 * score))]
-
-                if 'classifications' in detection:
-
-                    # To avoid duplicate colors with detection-only visualization, offset
-                    # the classification class index by the number of detection classes
-                    clss = TFDetector.NUM_DETECTOR_CATEGORIES + int(detection['classifications'][0][0])
-                    classifications = detection['classifications']
-                    if len(classifications) > max_classifications:
-                        classifications = classifications[0:max_classifications]
-                    for classification in classifications:
-                        p = classification[1]
-                        if p < classification_confidence_threshold:
-                            continue
-                        class_key = classification[0]
-                        if class_key in classification_label_map:
-                            class_name = classification_label_map[class_key]
-                        else:
-                            class_name = class_key
-                        displayed_label += ['{}: {:5.1%}'.format(class_name.lower(), classification[1])]
-
-                # ...if we have detection results
-                display_strs.append(displayed_label)
-                classes.append(clss)
-
-            # ...if the confidence of this detection is above threshold
-
-        # ...for each detection
-        display_boxes = np.array(display_boxes)
-
-        DetectorUtils.draw_bounding_boxes_on_image(image, display_boxes, classes,
-                                                   display_strs=display_strs, thickness=thickness, expansion=expansion)
-
-    @staticmethod
-    def draw_bounding_boxes_on_image(image,
-                                     boxes,
-                                     classes,
-                                     thickness=4,
-                                     expansion=0,
-                                     display_strs=()):
-        """
-        Draws bounding boxes on image.
-
-        Args:
-          image: a PIL.Image object.
-          boxes: a 2 dimensional numpy array of [N, 4]: (ymin, xmin, ymax, xmax).
-                 The coordinates are in normalized format between [0, 1].
-          classes: a list of ints or strings (that can be cast to ints) corresponding to the class labels of the boxes.
-                 This is only used for selecting the color to render the bounding box in.
-          thickness: line thickness in pixels. Default value is 4.
-          expansion: number of pixels to expand bounding boxes on each side.  Default is 0.
-          display_strs: list of list of strings.
-                                 a list of strings for each bounding box.
-                                 The reason to pass a list of strings for a
-                                 bounding box is that it might contain
-                                 multiple labels.
-        """
-
-        boxes_shape = boxes.shape
-        if not boxes_shape:
-            return
-        if len(boxes_shape) != 2 or boxes_shape[1] != 4:
-            # print('Input must be of size [N, 4], but is ' + str(boxes_shape))
-            return  # no object detection on this image, return
-        for i in range(boxes_shape[0]):
-            if display_strs:
-                display_str_list = display_strs[i]
-                DetectorUtils.draw_bounding_box_on_image(image,
-                                                         boxes[i, 0], boxes[i, 1], boxes[i, 2], boxes[i, 3],
-                                                         classes[i],
-                                                         thickness=thickness, expansion=expansion,
-                                                         display_str_list=display_str_list)
-
-    @staticmethod
-    def draw_bounding_box_on_image(image,
-                                   ymin,
-                                   xmin,
-                                   ymax,
-                                   xmax,
-                                   clss=None,
-                                   thickness=4,
-                                   expansion=0,
-                                   display_str_list=(),
-                                   use_normalized_coordinates=True,
-                                   label_font_size=16):
-        """
-        Adds a bounding box to an image.
-
-        Bounding box coordinates can be specified in either absolute (pixel) or
-        normalized coordinates by setting the use_normalized_coordinates argument.
-
-        Each string in display_str_list is displayed on a separate line above the
-        bounding box in black text on a rectangle filled with the input 'color'.
-        If the top of the bounding box extends to the edge of the image, the strings
-        are displayed below the bounding box.
-
-        Args:
-        image: a PIL.Image object.
-        ymin: ymin of bounding box - upper left.
-        xmin: xmin of bounding box.
-        ymax: ymax of bounding box.
-        xmax: xmax of bounding box.
-        clss: str, the class of the object in this bounding box - will be cast to an int.
-        thickness: line thickness. Default value is 4.
-        expansion: number of pixels to expand bounding boxes on each side.  Default is 0.
-        display_str_list: list of strings to display in box
-            (each to be shown on its own line).
-            use_normalized_coordinates: If True (default), treat coordinates
-            ymin, xmin, ymax, xmax as relative to the image.  Otherwise treat
-            coordinates as absolute.
-        label_font_size: font size to attempt to load arial.ttf with
-        """
-        if clss is None:
-            color = DetectorUtils.COLORS[1]
-        else:
-            color = DetectorUtils.COLORS[int(clss) % len(DetectorUtils.COLORS)]
-
-        draw = ImageDraw.Draw(image)
-        im_width, im_height = image.size
-        if use_normalized_coordinates:
-            (left, right, top, bottom) = (xmin * im_width, xmax * im_width,
-                                          ymin * im_height, ymax * im_height)
-        else:
-            (left, right, top, bottom) = (xmin, xmax, ymin, ymax)
-
-        if expansion > 0:
-            left -= expansion
-            right += expansion
-            top -= expansion
-            bottom += expansion
-
-        draw.line([(left, top), (left, bottom), (right, bottom),
-                   (right, top), (left, top)], width=thickness, fill=color)
-
-        try:
-            font = ImageFont.truetype('arial.ttf', label_font_size)
-        except IOError:
-            font = ImageFont.load_default()
-
-        # If the total height of the display strings added to the top of the bounding
-        # box exceeds the top of the image, stack the strings below the bounding box
-        # instead of above.
-        display_str_heights = [font.getsize(ds)[1] for ds in display_str_list]
-
-        # Each display_str has a top and bottom margin of 0.05x.
-        total_display_str_height = (1 + 2 * 0.05) * sum(display_str_heights)
-
-        if top > total_display_str_height:
-            text_bottom = top
-        else:
-            text_bottom = bottom + total_display_str_height
-
-        # Reverse list and print from bottom to top.
-        for display_str in display_str_list[::-1]:
-            text_width, text_height = font.getsize(display_str)
-            margin = np.ceil(0.05 * text_height)
-
-            draw.rectangle(
-                [(left, text_bottom - text_height - 2 * margin), (left + text_width,
-                                                                  text_bottom)],
-                fill=color)
-
-            draw.text(
-                (left + margin, text_bottom - text_height - margin),
-                display_str,
-                fill='black',
-                font=font)
-
-            text_bottom -= (text_height + 2 * margin)
-
-
 class TFDetector:
     """
     A detector model loaded at the time of initialization. It is intended to be used with
@@ -490,7 +130,7 @@ class TFDetector:
     FAILURE_IMAGE_OPEN = 'Failure image access'
 
     DEFAULT_RENDERING_CONFIDENCE_THRESHOLD = 0.85  # to render bounding boxes
-    DEFAULT_OUTPUT_CONFIDENCE_THRESHOLD = 0.3  # to include in the output json file
+    DEFAULT_OUTPUT_CONFIDENCE_THRESHOLD = 0.1  # to include in the output json file
 
     DEFAULT_DETECTOR_LABEL_MAP = {
         '1': 'animal',
@@ -512,8 +152,12 @@ class TFDetector:
         self.class_tensor = detection_graph.get_tensor_by_name('detection_classes:0')
 
     @staticmethod
+    def round_and_make_float(d, precision=4):
+        return truncate_float(float(d), precision=precision)
+
+    @staticmethod
     def __convert_coords(np_array):
-        """ Two functionalities: convert the numpy floats to Python floats, and also change the coordinates from
+        """ Two effects: convert the numpy floats to Python floats, and also change the coordinates from
         [y1, x1, y2, x2] to [x1, y1, width_box, height_box] (in relative coordinates still).
 
         Args:
@@ -530,7 +174,7 @@ class TFDetector:
 
         # convert numpy floats to Python floats
         for i, d in enumerate(new):
-            new[i] = DetectorUtils.round_and_make_float(d, precision=TFDetector.COORD_DIGITS)
+            new[i] = TFDetector.round_and_make_float(d, precision=TFDetector.COORD_DIGITS)
         return new
 
     @staticmethod
@@ -601,16 +245,16 @@ class TFDetector:
                 if s > detection_threshold:
                     detection_entry = {
                         'category': str(int(c)),  # use string type for the numerical class label, not int
-                        'conf': DetectorUtils.truncate_float(float(s),  # cast to float for json serialization
-                                                             precision=TFDetector.CONF_DIGITS),
+                        'conf': truncate_float(float(s),  # cast to float for json serialization
+                                               precision=TFDetector.CONF_DIGITS),
                         'bbox': TFDetector.__convert_coords(b)
                     }
                     detections_cur_image.append(detection_entry)
                     if s > max_detection_conf:
                         max_detection_conf = s
 
-            result['max_detection_conf'] = DetectorUtils.truncate_float(float(max_detection_conf),
-                                                                        precision=TFDetector.CONF_DIGITS)
+            result['max_detection_conf'] = truncate_float(float(max_detection_conf),
+                                                          precision=TFDetector.CONF_DIGITS)
             result['detections'] = detections_cur_image
 
         except Exception as e:
@@ -646,7 +290,7 @@ def load_and_run_detector(model_file, image_file_names, output_dir,
         try:
             start_time = time.time()
 
-            image = DetectorUtils.load_image(im_file)
+            image = viz_utils.load_image(im_file)
 
             elapsed = time.time() - start_time
             time_load.append(elapsed)
@@ -675,9 +319,9 @@ def load_and_run_detector(model_file, image_file_names, output_dir,
 
         try:
             # image is modified in place
-            DetectorUtils.render_detection_bounding_boxes(result['detections'], image,
-                                                          label_map=TFDetector.DEFAULT_DETECTOR_LABEL_MAP,
-                                                          confidence_threshold=render_confidence_threshold)
+            viz_utils.render_detection_bounding_boxes(result['detections'], image,
+                                                      label_map=TFDetector.DEFAULT_DETECTOR_LABEL_MAP,
+                                                      confidence_threshold=render_confidence_threshold)
             fn = os.path.basename(im_file).lower()
             name, ext = os.path.splitext(fn)
             fn = '{}{}{}'.format(name, ImagePathUtils.DETECTION_FILENAME_INSERT, '.jpg')  # save all as JPG
@@ -741,7 +385,8 @@ def main():
         '--threshold',
         type=float,
         default=TFDetector.DEFAULT_RENDERING_CONFIDENCE_THRESHOLD,
-        help='Confidence threshold between 0 and 1.0; only render boxes above this confidence (but only boxes above 0.3 confidence will be considered at all)'
+        help=('Confidence threshold between 0 and 1.0; only render boxes above this confidence'
+              ' (but only boxes above 0.1 confidence will be considered at all)')
     )
 
     if len(sys.argv[1:]) == 0:
