+++ visualization/visualization_utils.py
@@ -1,49 +1,60 @@
-#####
-#
-# visualization_utils.py
-#
-# Core rendering functions shared across visualization scripts
-#
-#####
+"""
+visualization_utils.py
+
+Core rendering functions shared across visualization scripts
+"""
 
 #%% Constants and imports
 
 import matplotlib.pyplot as plt
 import matplotlib.cm as cm
 import numpy as np
-from PIL import Image, ImageFile
-from detection.run_tf_detector import DetectorUtils
+from PIL import Image, ImageFile, ImageFont, ImageDraw
 
 from data_management.annotations import annotation_constants
 
 ImageFile.LOAD_TRUNCATED_IMAGES = True
 
 
-#%% Classes
-
-
 #%% Functions
 
-def open_image(input):
+def open_image(input_file):
     """
     Opens an image in binary format using PIL.Image and convert to RGB mode. This operation is lazy; image will
     not be actually loaded until the first operation that needs to load it (for example, resizing), so file opening
     errors can show up later.
 
     Args:
-        input: an image in binary format read from the POST request's body or
+        input_file: an image in binary format read from the POST request's body or
             path to an image file (anything that PIL can open)
 
     Returns:
         an PIL image object in RGB mode
     """
 
-    image = Image.open(input)
+    image = Image.open(input_file)
     if image.mode not in ('RGBA', 'RGB', 'L'):
-        raise AttributeError('Input image {} uses unsupported mode {}'.format(input,image.mode))
+        raise AttributeError('Input image {} uses unsupported mode {}'.format(input_file, image.mode))
     if image.mode == 'RGBA' or image.mode == 'L':
         # PIL.Image.convert() returns a converted copy of this image
         image = image.convert(mode='RGB')
+    return image
+
+
+def load_image(input_file):
+    """
+    Loads the image at input_file as a PIL Image into memory; Image.open() used in open_image() is lazy and
+    errors will occur downstream if not explicitly loaded
+
+    Args:
+        input_file: an image in binary format read from the POST request's body or
+            path to an image file (anything that PIL can open)
+
+    Returns:
+        an PIL image object in RGB mode
+    """
+    image = open_image(input_file)
+    image.load()
     return image
 
 
@@ -93,6 +104,291 @@ def show_images_in_a_row(images):
     return fig
 
 
+# The following three functions are modified versions of those at:
+# https://github.com/tensorflow/models/blob/master/research/object_detection/utils/visualization_utils.py
+
+COLORS = [
+        'AliceBlue', 'Red', 'RoyalBlue', 'Gold', 'Chartreuse', 'Aqua', 'Azure',
+        'Beige', 'Bisque', 'BlanchedAlmond', 'BlueViolet', 'BurlyWood', 'CadetBlue',
+        'AntiqueWhite', 'Chocolate', 'Coral', 'CornflowerBlue', 'Cornsilk', 'Crimson',
+        'Cyan', 'DarkCyan', 'DarkGoldenRod', 'DarkGrey', 'DarkKhaki', 'DarkOrange',
+        'DarkOrchid', 'DarkSalmon', 'DarkSeaGreen', 'DarkTurquoise', 'DarkViolet',
+        'DeepPink', 'DeepSkyBlue', 'DodgerBlue', 'FireBrick', 'FloralWhite',
+        'ForestGreen', 'Fuchsia', 'Gainsboro', 'GhostWhite', 'GoldenRod',
+        'Salmon', 'Tan', 'HoneyDew', 'HotPink', 'IndianRed', 'Ivory', 'Khaki',
+        'Lavender', 'LavenderBlush', 'LawnGreen', 'LemonChiffon', 'LightBlue',
+        'LightCoral', 'LightCyan', 'LightGoldenRodYellow', 'LightGray', 'LightGrey',
+        'LightGreen', 'LightPink', 'LightSalmon', 'LightSeaGreen', 'LightSkyBlue',
+        'LightSlateGray', 'LightSlateGrey', 'LightSteelBlue', 'LightYellow', 'Lime',
+        'LimeGreen', 'Linen', 'Magenta', 'MediumAquaMarine', 'MediumOrchid',
+        'MediumPurple', 'MediumSeaGreen', 'MediumSlateBlue', 'MediumSpringGreen',
+        'MediumTurquoise', 'MediumVioletRed', 'MintCream', 'MistyRose', 'Moccasin',
+        'NavajoWhite', 'OldLace', 'Olive', 'OliveDrab', 'Orange', 'OrangeRed',
+        'Orchid', 'PaleGoldenRod', 'PaleGreen', 'PaleTurquoise', 'PaleVioletRed',
+        'PapayaWhip', 'PeachPuff', 'Peru', 'Pink', 'Plum', 'PowderBlue', 'Purple',
+        'RosyBrown', 'Aquamarine', 'SaddleBrown', 'Green', 'SandyBrown',
+        'SeaGreen', 'SeaShell', 'Sienna', 'Silver', 'SkyBlue', 'SlateBlue',
+        'SlateGray', 'SlateGrey', 'Snow', 'SpringGreen', 'SteelBlue', 'GreenYellow',
+        'Teal', 'Thistle', 'Tomato', 'Turquoise', 'Violet', 'Wheat', 'White',
+        'WhiteSmoke', 'Yellow', 'YellowGreen'
+    ]
+
+
+def render_detection_bounding_boxes(detections, image,
+                                    label_map={},
+                                    classification_label_map={},
+                                    confidence_threshold=0.8, thickness=4, expansion=0,
+                                    classification_confidence_threshold=0.3,
+                                    max_classifications=3):
+    """
+    Renders bounding boxes, label, and confidence on an image if confidence is above the threshold.
+
+    This works with the output of the batch processing API.
+
+    Supports classification, if the detection contains classification results according to the
+    API output version 1.0.
+
+    Args:
+
+        detections: detections on the image, example content:
+            [
+                {
+                    "category": "2",
+                    "conf": 0.996,
+                    "bbox": [
+                        0.0,
+                        0.2762,
+                        0.1234,
+                        0.2458
+                    ]
+                }
+            ]
+
+            ...where the bbox coordinates are [x, y, box_width, box_height].
+
+            (0, 0) is the upper-left.  Coordinates are normalized.
+
+            Supports classification results, if *detections* have the format
+            [
+                {
+                    "category": "2",
+                    "conf": 0.996,
+                    "bbox": [
+                        0.0,
+                        0.2762,
+                        0.1234,
+                        0.2458
+                    ]
+                    "classifications": [
+                        ["3", 0.901],
+                        ["1", 0.071],
+                        ["4", 0.025]
+                    ]
+                }
+            ]
+
+        image: PIL.Image object, output of generate_detections.
+
+        label_map: optional, mapping the numerical label to a string name. The type of the numerical label
+            (default string) needs to be consistent with the keys in label_map; no casting is carried out.
+
+        classification_label_map: optional, mapping of the string class labels to the actual class names.
+            The type of the numerical label (default string) needs to be consistent with the keys in
+            label_map; no casting is carried out.
+
+        confidence_threshold: optional, threshold above which the bounding box is rendered.
+        thickness: line thickness in pixels. Default value is 4.
+        expansion: number of pixels to expand bounding boxes on each side.  Default is 0.
+        classification_confidence_threshold: confidence above which classification result is retained.
+        max_classifications: maximum number of classification results retained for one image.
+
+    image is modified in place.
+    """
+
+    display_boxes = []
+    display_strs = []  # list of lists, one list of strings for each bounding box (to accommodate multiple labels)
+    classes = []  # for color selection
+
+    for detection in detections:
+
+        score = detection['conf']
+        if score >= confidence_threshold:
+
+            x1, y1, w_box, h_box = detection['bbox']
+            display_boxes.append([y1, x1, y1 + h_box, x1 + w_box])
+            clss = detection['category']
+            label = label_map[clss] if clss in label_map else clss
+            displayed_label = ['{}: {}%'.format(label, round(100 * score))]
+
+            if 'classifications' in detection:
+
+                # To avoid duplicate colors with detection-only visualization, offset
+                # the classification class index by the number of detection classes
+                clss = annotation_constants.NUM_DETECTOR_CATEGORIES + int(detection['classifications'][0][0])
+                classifications = detection['classifications']
+                if len(classifications) > max_classifications:
+                    classifications = classifications[0:max_classifications]
+                for classification in classifications:
+                    p = classification[1]
+                    if p < classification_confidence_threshold:
+                        continue
+                    class_key = classification[0]
+                    if class_key in classification_label_map:
+                        class_name = classification_label_map[class_key]
+                    else:
+                        class_name = class_key
+                    displayed_label += ['{}: {:5.1%}'.format(class_name.lower(), classification[1])]
+
+            # ...if we have detection results
+            display_strs.append(displayed_label)
+            classes.append(clss)
+
+        # ...if the confidence of this detection is above threshold
+
+    # ...for each detection
+    display_boxes = np.array(display_boxes)
+
+    draw_bounding_boxes_on_image(image, display_boxes, classes,
+                                 display_strs=display_strs, thickness=thickness, expansion=expansion)
+
+
+def draw_bounding_boxes_on_image(image,
+                                 boxes,
+                                 classes,
+                                 thickness=4,
+                                 expansion=0,
+                                 display_strs=()):
+    """
+    Draws bounding boxes on image.
+
+    Args:
+      image: a PIL.Image object.
+      boxes: a 2 dimensional numpy array of [N, 4]: (ymin, xmin, ymax, xmax).
+             The coordinates are in normalized format between [0, 1].
+      classes: a list of ints or strings (that can be cast to ints) corresponding to the class labels of the boxes.
+             This is only used for selecting the color to render the bounding box in.
+      thickness: line thickness in pixels. Default value is 4.
+      expansion: number of pixels to expand bounding boxes on each side.  Default is 0.
+      display_strs: list of list of strings.
+                             a list of strings for each bounding box.
+                             The reason to pass a list of strings for a
+                             bounding box is that it might contain
+                             multiple labels.
+    """
+
+    boxes_shape = boxes.shape
+    if not boxes_shape:
+        return
+    if len(boxes_shape) != 2 or boxes_shape[1] != 4:
+        # print('Input must be of size [N, 4], but is ' + str(boxes_shape))
+        return  # no object detection on this image, return
+    for i in range(boxes_shape[0]):
+        if display_strs:
+            display_str_list = display_strs[i]
+            draw_bounding_box_on_image(image,
+                                       boxes[i, 0], boxes[i, 1], boxes[i, 2], boxes[i, 3],
+                                       classes[i],
+                                       thickness=thickness, expansion=expansion,
+                                       display_str_list=display_str_list)
+
+
+def draw_bounding_box_on_image(image,
+                               ymin,
+                               xmin,
+                               ymax,
+                               xmax,
+                               clss=None,
+                               thickness=4,
+                               expansion=0,
+                               display_str_list=(),
+                               use_normalized_coordinates=True,
+                               label_font_size=16):
+    """
+    Adds a bounding box to an image.
+
+    Bounding box coordinates can be specified in either absolute (pixel) or
+    normalized coordinates by setting the use_normalized_coordinates argument.
+
+    Each string in display_str_list is displayed on a separate line above the
+    bounding box in black text on a rectangle filled with the input 'color'.
+    If the top of the bounding box extends to the edge of the image, the strings
+    are displayed below the bounding box.
+
+    Args:
+    image: a PIL.Image object.
+    ymin: ymin of bounding box - upper left.
+    xmin: xmin of bounding box.
+    ymax: ymax of bounding box.
+    xmax: xmax of bounding box.
+    clss: str, the class of the object in this bounding box - will be cast to an int.
+    thickness: line thickness. Default value is 4.
+    expansion: number of pixels to expand bounding boxes on each side.  Default is 0.
+    display_str_list: list of strings to display in box
+        (each to be shown on its own line).
+        use_normalized_coordinates: If True (default), treat coordinates
+        ymin, xmin, ymax, xmax as relative to the image.  Otherwise treat
+        coordinates as absolute.
+    label_font_size: font size to attempt to load arial.ttf with
+    """
+    if clss is None:
+        color = COLORS[1]
+    else:
+        color = COLORS[int(clss) % len(COLORS)]
+
+    draw = ImageDraw.Draw(image)
+    im_width, im_height = image.size
+    if use_normalized_coordinates:
+        (left, right, top, bottom) = (xmin * im_width, xmax * im_width,
+                                      ymin * im_height, ymax * im_height)
+    else:
+        (left, right, top, bottom) = (xmin, xmax, ymin, ymax)
+
+    if expansion > 0:
+        left -= expansion
+        right += expansion
+        top -= expansion
+        bottom += expansion
+
+    draw.line([(left, top), (left, bottom), (right, bottom),
+               (right, top), (left, top)], width=thickness, fill=color)
+
+    try:
+        font = ImageFont.truetype('arial.ttf', label_font_size)
+    except IOError:
+        font = ImageFont.load_default()
+
+    # If the total height of the display strings added to the top of the bounding
+    # box exceeds the top of the image, stack the strings below the bounding box
+    # instead of above.
+    display_str_heights = [font.getsize(ds)[1] for ds in display_str_list]
+
+    # Each display_str has a top and bottom margin of 0.05x.
+    total_display_str_height = (1 + 2 * 0.05) * sum(display_str_heights)
+
+    if top > total_display_str_height:
+        text_bottom = top
+    else:
+        text_bottom = bottom + total_display_str_height
+
+    # Reverse list and print from bottom to top.
+    for display_str in display_str_list[::-1]:
+        text_width, text_height = font.getsize(display_str)
+        margin = np.ceil(0.05 * text_height)
+
+        draw.rectangle(
+            [(left, text_bottom - text_height - 2 * margin), (left + text_width,
+                                                              text_bottom)],
+            fill=color)
+
+        draw.text(
+            (left + margin, text_bottom - text_height - margin),
+            display_str,
+            fill='black',
+            font=font)
+
+        text_bottom -= (text_height + 2 * margin)
+
+
 def render_iMerit_boxes(boxes, classes, image,
                         label_map=annotation_constants.bbox_category_id_to_name):
     """
@@ -100,6 +396,7 @@ def render_iMerit_boxes(boxes, classes, image,
 
     Args:
         boxes: bounding box annotations from iMerit, format is [x_rel, y_rel, w_rel, h_rel] (rel = relative coords)
+        classes: the class IDs of the predicted class of each box/object
         image: PIL.Image object to annotate on
         label_map: optional dict mapping classes to a string for display
 
@@ -122,7 +419,7 @@ def render_iMerit_boxes(boxes, classes, image,
         display_strs.append([clss])
 
     display_boxes = np.array(display_boxes)
-    DetectorUtils.draw_bounding_boxes_on_image(image, display_boxes, classes, display_strs=display_strs)
+    draw_bounding_boxes_on_image(image, display_boxes, classes, display_strs=display_strs)
 
 
 def render_megadb_bounding_boxes(boxes_info, image):
@@ -152,7 +449,7 @@ def render_megadb_bounding_boxes(boxes_info, image):
         classes.append(annotation_constants.bbox_category_name_to_id[b['category']])
 
     display_boxes = np.array(display_boxes)
-    DetectorUtils.draw_bounding_boxes_on_image(image, display_boxes, classes, display_strs=display_strs)
+    draw_bounding_boxes_on_image(image, display_boxes, classes, display_strs=display_strs)
 
 
 def render_db_bounding_boxes(boxes, classes, image, original_size=None,
@@ -190,25 +487,8 @@ def render_db_bounding_boxes(boxes, classes, image, original_size=None,
         display_strs.append([str(clss)])  # need to be a string here because PIL needs to iterate through chars
 
     display_boxes = np.array(display_boxes)
-    DetectorUtils.draw_bounding_boxes_on_image(image, display_boxes, classes, display_strs=display_strs,
-                                 thickness=thickness,expansion=expansion)
-
-
-def render_detection_bounding_boxes(detections, image,
-                                        label_map={},
-                                        classification_label_map={},
-                                        confidence_threshold=0.8, thickness=4, expansion=0,
-                                        classification_confidence_threshold=0.3,
-                                        max_classifications=3):
-    """
-    Renders bounding boxes, label, and confidence on an image if confidence is above the threshold.
-    
-    Simple wrapper around the corresponding DetectorUtils function.
-    """
-    DetectorUtils.render_detection_bounding_boxes(detections,image,label_map,classification_label_map,
-                                                  confidence_threshold,thickness,expansion,
-                                                  classification_confidence_threshold,
-                                                  max_classifications)
+    draw_bounding_boxes_on_image(image, display_boxes, classes, display_strs=display_strs,
+                                 thickness=thickness, expansion=expansion)
     
     
 def plot_confusion_matrix(matrix, classes,
@@ -217,7 +497,7 @@ def plot_confusion_matrix(matrix, classes,
                           cmap=plt.cm.Blues,
                           vmax=None,
                           use_colorbar=True,
-                          y_label = True):
+                          y_label=True):
 
     """
     This function plots a confusion matrix.
@@ -245,10 +525,11 @@ def plot_confusion_matrix(matrix, classes,
         matrix = matrix.astype(np.double) / (matrix.sum(axis=1, keepdims=True) + 1e-7)
 
     plt.imshow(matrix, interpolation='nearest', cmap=cmap, vmax=vmax)
-    plt.title(title) #,fontsize=22)
+    plt.title(title)   # ,fontsize=22)
 
     if use_colorbar:
-        plt.colorbar(fraction=0.046, pad=0.04, ticks=[0.0,0.25,0.5,0.75,1.0]).set_ticklabels(['0%','25%','50%','75%','100%'])
+        plt.colorbar(fraction=0.046, pad=0.04,
+                     ticks=[0.0,0.25,0.5,0.75,1.0]).set_ticklabels(['0%','25%','50%','75%','100%'])
 
     tick_marks = np.arange(len(classes))
     plt.xticks(tick_marks, classes, rotation=90)
@@ -294,8 +575,7 @@ def plot_precision_recall_curve(precisions, recalls, title='Precision/Recall cur
     step_kwargs = ({'step': 'post'})
     fig = plt.figure()
     plt.title(title)
-    plt.step(recalls, precisions, color='b', alpha=0.2,
-                where='post')
+    plt.step(recalls, precisions, color='b', alpha=0.2, where='post')
     plt.fill_between(recalls, precisions, alpha=0.2, color='b', **step_kwargs)
 
     plt.xlabel('Recall')
